Split out tscclock into virtio-specific module
tscclock is not required for ukvm, split it and dependencies into a
separate module built only for the virtio target. No functional change
intended.
