+++ kernel/virtio/clock_subr.h
